+++ src/egeo-demo/barrels.ts
@@ -30,6 +30,7 @@ import { StItemListDemoModule } from './st-item-list-demo/st-item-list-demo.modu
 import { StLabelDemoModule } from './st-label-demo/st-label-demo.module';
 import { StLauncherDemoModule } from './st-launcher-demo/st-launcher-demo.module';
 import { StModalDemoModule } from './st-modal-demo/st-modal-demo.module';
+import { StModal2DemoModule } from './st-modal2-demo/st-modal2-demo.module';
 import { StPageTitleDemoModule } from './st-page-title-demo/st-page-title-demo.module';
 import { StPaginationDemoModule } from './st-pagination-demo/st-pagination-demo.module';
 import { StPopOverDemoModule } from './st-pop-over-demo/st-pop-over-demo.module';
@@ -78,6 +79,7 @@ export const EGEO_DEMO_MODULES: any[] = [
    StLabelDemoModule,
    StLauncherDemoModule,
    StModalDemoModule,
+   StModal2DemoModule,
    StPageTitleDemoModule,
    StPaginationDemoModule,
    StPopOverDemoModule,